+++ src/components.ts
@@ -1,0 +1,2 @@
+export * from './grid/base/grid';
+export * from './pager/pager';