+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## [Unreleased]
 
+## 16.1.35 (2018-04-17)
+
+### Grid
+#### Bug Fixes
+- Maximum call stack issue while filtering date column with `disablePageWiseAggregates` is resolved.
+- Provided locale support for custom filter labels and column chooser.
+- Between operator provides incorrect result in Excel filter issue fixed.
+- Content rendering delayed while using virtual scrolling with grouping when scrolling horizontally is resolved.
+#### Breaking Changes
+- The Locale properties such as `OK`, `Filter`, `Clear` are removed instead use `OKButton`, `FilterButton`, `ClearButton`.
+
 ## 16.1.34 (2018-04-10)
 
 ### Grid